multivariate normal torch experiment
